putbase can print long long int

+++ Printf/includes/libftprintf.h
@@ -6,14 +6,15 @@
 /*   By: nsainton <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/11/14 16:25:56 by nsainton          #+#    #+#             */
-/*   Updated: 2022/11/14 17:44:59 by nsainton         ###   ########.fr       */
+/*   Updated: 2022/11/14 20:34:06 by nsainton         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #ifndef LIBFT_PRINTF_H
 # define LIBFT_PRINTF_H
 # include "../Libft/libft.h"
+# include <limits.h>
 
-void	ft_putbase(int nbr, int maj, const char *base);
+void	ft_putbase(long long int nbr, int maj, const char *base);
 
 #endif
